Handle blocks out of topological order
This commit refactors the block scanning algorithm to handle checking for
relevant transactions when the block is not in topological order (for example,
when CTOR is active).

We do this by keeping a map of all of the inputs in a block. Whenever a tx
is processed, we check its hash against the input map, if there is a match
we process that previous transaction again and likewise recursively process
its dependencies.