Use custom DNS resolver for neutrino
Because neutrino will likely be running in a mobile environment we want to
make use of orbot's DNS proxy if it's available so that the wallet can be out
of the box compatible with orbot.

To that end this new custom DNS resolution function tries a name resolution
on localhost first before defaulting to the dns resolver in the net package.