Refactor address handling
This commit refactors the address handling so that it generates a buffer of
20 unused addresses on wallet creation. Each time a new transaction is ingested
it will maybe extend the keychain to try to maintain that buffer of unused address.

This permits the CurrentAddress API call to work as read only as opposed to
read and maybe write a new address to the db which which blocks during initial
chain sync and rescan.