Create wallet with customizable birthday
This commit will use the genesis block as the wallet birthday if the user
entered a seed via the command line prompt. It also adds a birthday timestamp
to the CreateWallet gRPC API request so that clients can specify the wallet
birthday when creating a wallet. This is would most likely be used when
restoring from seed.